Fix SonarCloud issues: Replace wildcards with specific types and extract string literals as constants

+++ src/main/java/com/loadbalancer/model/dto/HealthStatus.java
@@ -15,5 +15,5 @@ public class HealthStatus {
   private int activeConnections;
   private LocalDateTime lastChecked;
   private String statusMessage;
-  private long responseTime;
+  private double responseTime;
 }
